+++ lib/envalid.rb
@@ -1,8 +1,0 @@
-# frozen_string_literal: true
-
-require_relative "envalid/version"
-
-module Envalid
-  class Error < StandardError; end
-  # Your code goes here...
-end