+++ spec/envalit_spec.rb
@@ -1,0 +1,13 @@
+# frozen_string_literal: true
+
+require "spec_helper"
+
+RSpec.describe Envalit do
+  it "has a version number" do
+    expect(Envalit::VERSION).not_to be nil
+  end
+
+  it "exposes the Loader class" do
+    expect(Envalit::Loader).to be_a(Class)
+  end
+end