+++ spec/spec_helper.rb
@@ -1,7 +1,9 @@
 # frozen_string_literal: true
 
-require "envalid"
-
+require "envalit"
+require "tmpdir"
+require "fileutils"
+require "pry"
 RSpec.configure do |config|
   # Enable flags like --only-failures and --next-failure
   config.example_status_persistence_file_path = ".rspec_status"